Refactored to reflect chapter reorg

+++ docker-images/pulsar-standalone/manning/security/publish-credentials.sh
@@ -1,10 +1,21 @@
 #!/bin/bash
 
-# Share the tokens with the client     
-cp /pulsar/manning/security/jwt/*-token.txt /pulsar/manning/dropbox/  
+# Share the client X.509 certificates
+if compgen -G "/pulsar/manning/security/authentication/tls/*.pem" > /dev/null; then
+   cp /pulsar/manning/security/authentication/tls/*.pem /pulsar/manning/dropbox/ 
+fi
+
+# Share the JWT tokens with the client     
+if compgen -G "/pulsar/manning/security/authentication/jwt/*-token.txt" > /dev/null; then
+   cp /pulsar/manning/security/authentication/jwt/*-token.txt /pulsar/manning/dropbox/  
+fi
 
 # Share the public encryption key
-cp /pulsar/manning/security/encryption/*.pem /pulsar/manning/dropbox/ 
+if compgen -G "/pulsar/manning/security/encryption/*.pem" > /dev/null; then
+   cp /pulsar/manning/security/encryption/*.pem /pulsar/manning/dropbox/
+fi
 
 # Share the public CA Truststore
-cp /pulsar/manning/security/cert-authority/certs/ca.cert.pem /pulsar/manning/dropbox/
+if [ -f "/pulsar/manning/security/cert-authority/certs/ca.cert.pem" ]; then
+   cp /pulsar/manning/security/cert-authority/certs/ca.cert.pem /pulsar/manning/dropbox/
+fi
